fixes: - app-nav-bar uses the repo pattern: overrided -> componentName - Removed domain-workflows-basic-lazy.tsx; domain-workflows dynamically imports the basic workflows component directly - Moved gate components into same-name folders
