--- src/views/domain-workflows/domain-workflows-basic-lazy.tsx
@@ -1,8 +1,0 @@
-'use client';
-import dynamic from 'next/dynamic';
-
-const DomainWorkflowsBasic = dynamic(
-  () => import('@/views/domain-workflows-basic/domain-workflows-basic')
-);
-
-export default DomainWorkflowsBasic;
